Move imports to src directory; add README.md
Restructures the directory so that the imported files are contained in
the src directory to keep things cleaner.

Also adds a README.md file with some basic installation and exeuction
instructions.
