+++ exploration/grid-poc/README.md
@@ -1,0 +1,25 @@
+# Plantworld
+
+Prototype Plantworld cell-based grid world
+
+## Installation
+
+Install the Python packages as listed in teh `requirements.txt` file:
+```
+$ python3 -m venv venv
+$ . ./venv/bin/activate
+% pip install pip --upgrace
+$ python3 -r requirements.txt
+```
+
+## Execution
+
+Run the `world.py` file as a script.
+```
+$ python3 ./world.py
+```
+
+The application will open a window and spend some time populating the world with a grid.
+Once completed some text will be displayed in the console requesting you to press ENTER to continue.
+Press ENTER and the Plantworld simulation will start running.
+
