+++ exploration/grid-poc/src/cells.py
